Navigation in android apps.
This app contains an order flow for cupcakes with options for quantity, flavor, and pickup date.
The order details get displayed on an order summary screen and can be shared to another app to
send the order.
Using Jetpack Compose with Kotlin, I understand the concept that how to use the Navigation component to build more complex apps with more screens and how to navigate and pass data between different composables.